figures out unique username and appears in chat
